backend/requirements.txt
# FastAPI and Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# OpenAI
openai>=1.10.0

# Qdrant Vector Database
qdrant-client>=1.7.1

# Environment variables
python-dotenv

# HTTP client
httpx

# Logging
structlog

# For handling markdown frontmatter
python-frontmatter
